fastAPI/requirements.txt
asyncpg==0.30.0
colorama==0.4.6
email_validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
loguru==0.7.3
numpy==2.2.4
pillow==11.1.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.1.0
redis==5.2.1
scipy==1.15.2
sentence-transformers==3.4.1
sympy==1.13.1
tokenizers==0.21.1
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tornado==6.4.2
tsidpy==1.1.5
typer==0.15.2
typing_extensions==4.12.2
uvicorn[standard]
python-multipart
python-jose[cryptography]
passlib[bcrypt]